fix(stack): declare diagnostics types for node10 resolution, and document it
`typesVersions` and `exports` describe the same subpaths to two different
resolvers, and only one of them is exercised here. `./diagnostics` was in
`exports` and not in `typesVersions`, so a consumer on classic node10
resolution — still the default when `module` is `commonjs` — resolved it to
`any` with no error at the import site. Nothing type-checks that: the
package builds, publishes and installs clean.

Guarded now rather than just fixed. The parity test asserts every exported
subpath has a `typesVersions` entry pointing at the same declaration file,
so the next subpath fails here instead of shipping half-declared. It is
keyed on what `exports` publishes, which leaves the reverse drift alone —
`secrets` has an entry for a subpath that no longer exists, dead weight
rather than a consumer-visible gap, and not this branch's to remove.

`skills/stash-encryption` carries the canonical subpath table and ships
inside the `stash` tarball. `adapter-kit` is listed there with a "do not
import in application code" caveat, so a tooling-only entry belongs in it;
`./diagnostics` is documented the same way. The root AGENTS.md list was
updated when the subpath landed, the skill was not — the map in AGENTS.md
points `packages/stack` subpath exports at this skill precisely so that
does not happen.

Also from review:

* `isPackageMissing` and `isSubpathUnavailable` were exported from
  `commands/doctor` solely so a test could reach them, against this repo's
  rule about testing through the public API. They are module-resolution
  error classifiers, so they now live in `module-error.ts` next to
  `isModuleNotFound` and `moduleNotFoundSpecifier` — imported by doctor,
  tested where they are defined, and no wider a surface than before.
* Counting failing rows by splitting on the message rather than
  `new RegExp(message)`. The needle is copy from `messages.ts`, and copy is
  free to grow a `.` or `(` that a regex would reinterpret instead of
  failing on.
